moved leaf api test case to evm transaction #146
